Add (and use) method to register custom attributes
